fix ef core to work and connect

+++ Models/OrderDetail.cs
@@ -9,9 +9,9 @@ namespace Restaurant.API.Models
     public class OrderDetail
     {
         [Key]
-        public long OrderDetailId { get; set; }
+        public int OrderDetailId { get; set; }
         
-        public long OrderMasterId { get; set; }
+        public int OrderMasterId { get; set; }
 
         public int FoodItemId { get; set; }
         public FoodItem FoodItem { get; set; }
